requirements.txt
matplotlib
numpy
pyrtlsdr
